Move payment details page objects for refactor
