+++ projects/magento/tests/BancontactCardPayment.spec.js
@@ -7,7 +7,7 @@ import {
   verifyFailedPayment,
   verifySuccessfulPayment,
 } from "../helpers/ScenarioHelper.js";
-import { PaymentDetailsPage } from "../pageObjects/checkout/PaymentDetails.page.js";
+import { PaymentDetailsPage } from "../pageObjects/plugin/PaymentDetails.page.js";
 
 const paymentResources = new PaymentResources();
 const bancontactCard = paymentResources.bcmc.be;
