+++ projects/magento/tests/GiftCardPayment.spec.js
@@ -8,7 +8,7 @@ import {
   verifyFailedPayment,
   verifySuccessfulPayment,
 } from "../helpers/ScenarioHelper.js";
-import { PaymentDetailsPage } from "../pageObjects/checkout/PaymentDetails.page.js";
+import { PaymentDetailsPage } from "../pageObjects/plugin/PaymentDetails.page.js";
 
 const paymentResources = new PaymentResources();
 const giftCard50EUR = paymentResources.giftCard.EUR50;
