+++ projects/magento/tests/IDealPayment.spec.js
@@ -1,6 +1,6 @@
 import { test } from "@playwright/test";
 import PaymentResources from "../../data/PaymentResources.js";
-import { PaymentDetailsPage } from "../pageObjects/checkout/PaymentDetails.page.js";
+import { PaymentDetailsPage } from "../pageObjects/plugin/PaymentDetails.page.js";
 import {
   goToShippingWithFullCart,
   proceedToPaymentAs,
