+++ projects/magento/tests/MultiBancoPayment.spec.js
@@ -6,7 +6,7 @@ import {
   verifySuccessfulPayment,
   verifyVoucherCouponGeneration,
 } from "../helpers/ScenarioHelper.js";
-import { PaymentDetailsPage } from "../pageObjects/checkout/PaymentDetails.page.js";
+import { PaymentDetailsPage } from "../pageObjects/plugin/PaymentDetails.page.js";
 
 const paymentResources = new PaymentResources();
 const user = paymentResources.guestUser.portuguese;
